Add package name validation to `init`

+++ src/commands/init/mod.rs
@@ -1,30 +1,36 @@
 use std::env;
-use std::path::PathBuf;
 
-use dialoguer::{theme::ColorfulTheme, Input};
+use dialoguer::{theme::ColorfulTheme, Confirm, Input};
 use eyre::Result;
 use node_semver::Version;
 use owo_colors::OwoColorize;
 use rust_i18n::t;
+use slug::slugify;
 use tokio::fs::File;
 use tokio::io::AsyncWriteExt;
 
 use crate::consts::NO_TEST_SPECIFIED;
 use crate::fnv_map;
-use crate::package_json::{PackageJson, Repository};
-use crate::util::{get_current_dir_name, input, validate_version};
+use crate::package_json::{validate_package_name, validate_version, PackageJson, Repository};
+use crate::util::{get_current_dir_name, input};
 
 mod default;
 
 pub async fn invoke(yes: bool) -> Result<()> {
     if yes {
         let default = default::invoke()?;
-        write_package_json(default).await?;
+        write_package_json(default, false).await?;
         return Ok(());
     }
 
     // Ask the questions
-    let name = input(&t!("package-name-prompt"), Some(get_current_dir_name()?))?;
+    let name = Input::with_theme(&ColorfulTheme::default())
+        .with_prompt(&t!("package-name-prompt"))
+        .default(slugify(get_current_dir_name()?))
+        .validate_with(validate_package_name)
+        .interact_text()?
+        .parse()?;
+
     let version: Version = Input::with_theme(&ColorfulTheme::default())
         .with_prompt(&t!("package-version-prompt"))
         .default("1.0.0".to_string())
@@ -36,7 +42,7 @@ pub async fn invoke(yes: bool) -> Result<()> {
     let test_command = input("Test command", Some(NO_TEST_SPECIFIED.into()))?;
 
     let mut git_repository = input(&t!("package-git-repository-prompt"), None)?;
-    if !git_repository.ends_with(".git") {
+    if !git_repository.is_empty() && !git_repository.ends_with(".git") {
         git_repository.push_str(".git");
     }
 
@@ -62,25 +68,36 @@ pub async fn invoke(yes: bool) -> Result<()> {
         },
     };
 
-    write_package_json(package_json).await?;
+    write_package_json(package_json, true).await?;
 
     Ok(())
 }
 
-async fn write_package_json(package_json: PackageJson) -> Result<PathBuf> {
+async fn write_package_json(package_json: PackageJson, ask_for_confirmation: bool) -> Result<()> {
     use serde_json::to_string_pretty;
 
-    let current_dir = env::current_dir()?;
-    let package_json_path = current_dir.join("package.json");
+    if ask_for_confirmation {
+        if Confirm::with_theme(&ColorfulTheme::default())
+            .with_prompt(&t!("write-package-json"))
+            .interact()?
+        {
+            let current_dir = env::current_dir()?;
+            let package_json_path = current_dir.join("package.json");
 
-    let mut file = File::create(&package_json_path).await?;
-    file.write_all(to_string_pretty(&package_json)?.as_bytes())
-        .await?;
+            let mut file = File::create(&package_json_path).await?;
 
-    println!(
-        "{} `{}`",
-        t!("successfully-wrote-package-json-to").green().bold(),
-        package_json_path.to_string_lossy()
-    );
-    Ok(package_json_path)
+            file.write_all(to_string_pretty(&package_json)?.as_bytes())
+                .await?;
+
+            println!(
+                "{} `{}`",
+                t!("successfully-wrote-package-json-to").green().bold(),
+                package_json_path.to_string_lossy()
+            );
+        } else {
+            eprintln!("{}", &t!("aborted-operation").red())
+        }
+    }
+
+    Ok(())
 }
